Create README.md
Abstract: 
This project is the final exam project in Digital Methods on Aarhus University, autumn 2021, and analyses lynching patterns between 1883-1941. 
The aim is to examine how the numbers of lynchings evolved over time, and to find what year that had the most lynchings and how the lynching patterns changed over time. 
The project is based on David Rigby's lynching data and data from the Tuskegee institute. Both datasets are present in this repository. 

This project compares date from the Tuskegee Institute with David Rigby's lynching data, to see if there is a notable difference. 
This project uses R to create visualizations of a detailed dataset of lynching to see the evolvement of the patterns in lynching in the pe-riod. 
The project found that the lynching peaked between 1883-1895, and 1884 had most lynching-cases. 
The data from Rigby and Tuskegee differs greatly in lynching numbers, and shows that Rigby only shared part of his used data. 
The project did not manage to create maps to show the patterns in lynching in the different states, however, it made me more aware of the importance of data tidying and preparation.